+++ tests/clients/test_alerts.py
@@ -62,20 +62,11 @@ class TestAlertsClient(object):
             "OPEN", self._alert_ids, reason=None
         )
 
-    def test_alerts_client_calls_update_state_with_pending_state_and_expected_value(
+    def test_alerts_client_calls_update_state_with_state_and_expected_value(
         self, mock_alerts_service, mock_alert_rules_service,
     ):
         alert_client = AlertsClient(mock_alerts_service, mock_alert_rules_service)
-        alert_client.pending(self._alert_ids)
+        alert_client.update_state("RESOLVED", self._alert_ids)
         mock_alerts_service.update_state.assert_called_once_with(
-            "PENDING", self._alert_ids, reason=None
-        )
-
-    def test_alerts_client_calls_update_state_with_in_progress_state_and_expected_value(
-        self, mock_alerts_service, mock_alert_rules_service,
-    ):
-        alert_client = AlertsClient(mock_alerts_service, mock_alert_rules_service)
-        alert_client.in_progress(self._alert_ids)
-        mock_alerts_service.update_state.assert_called_once_with(
-            "IN_PROGRESS", self._alert_ids, reason=None
+            "RESOLVED", self._alert_ids, reason=None
         )